+++ backend/uclapi/workspaces/views.py
@@ -5,8 +5,6 @@ from django.utils.decorators import method_decorator
 from rest_framework import generics
 from rest_framework.decorators import api_view
 from rest_framework.exceptions import ParseError
-from rest_framework.pagination import CursorPagination
-from rest_framework.response import Response
 
 from common.decorators import uclapi_protected_endpoint
 from common.helpers import PrettyJsonResponse as JsonResponse, pretty_response
@@ -16,6 +14,7 @@ from .models import Sensors, Historical, Surveys
 from .occupeye.api import OccupEyeApi
 from .occupeye.constants import OccupEyeConstants
 from .occupeye.exceptions import BadOccupEyeRequest
+from .pagination import HistoricalListCursorPagination
 from .serializers import SensorsSerializer, HistoricalSerializer, SurveysSerializer
 
 
@@ -379,20 +378,6 @@ class SensorsList(generics.ListAPIView):
         return pretty_response(super().list(request, *args, **kwargs), custom_header_data=kwargs)
 
 
-class HistoricalListCursorPagination(CursorPagination):
-    ordering = "datetime"
-
-    def get_paginated_response(self, data, **kwargs):
-        return Response({
-            "okay": True,
-            "data": {
-                "next": self.get_next_link(),
-                "previous": self.get_previous_link(),
-                "results": data
-            }
-        })
-
-
 class HistoricalList(generics.ListAPIView):
     queryset = Historical.objects.all().order_by("datetime")
     serializer_class = HistoricalSerializer
@@ -403,5 +388,5 @@ class HistoricalList(generics.ListAPIView):
     @method_decorator(uclapi_protected_endpoint(personal_data=False, last_modified_redis_key="Workspaces-Historical"))
     def list(self, request, *args, **kwargs):
         if self.request.query_params.get("survey_id", None) is None:
-            raise ParseError("Survey_id is a required field")
+            raise ParseError("survey_id is a required field")
         return pretty_response(super().list(request, *args, **kwargs), custom_header_data=kwargs)
